Removed progress bars from training to be better suited for use with 'sbatch'

+++ scripts/run_schedule.py
@@ -10,13 +10,15 @@ import datetime
 import argparse
 from multiprocessing import Process
 import time
-import logging
+from utils import parse_bool
 from tqdm import tqdm
+import random
 
 def get_run_schedule_args():
     parser = argparse.ArgumentParser()
     parser.add_argument('--storage_dir', type=str, required=True)
     parser.add_argument('--n_processes', type=int, default=1)
+    parser.add_argument('--pbar', type=parse_bool, default=True)
     return parser.parse_args()
 
 def run_schedule(serial_args, process_i=0):
@@ -34,30 +36,45 @@ def run_schedule(serial_args, process_i=0):
 
         return all_unhatched_seeds
 
-    all_unhatched_seeds = get_unhatched_seeds(storage_dir)
+    process_logger = create_logger(f"process_{process_i}", save_dir=storage_dir)
 
     # For each unhatched experiments, load the args and try to train the model
     call_i = 0
-    while len(all_unhatched_seeds) > 0:
+    while True:
         call_i += 1
         try:
+            # random sleep time to delay different processes from one another
+            time.sleep(random.uniform(0, 5))
+            
             all_unhatched_seeds = get_unhatched_seeds(storage_dir)
-            seed_dir = all_unhatched_seeds[0]
-            os.remove(str(seed_dir / 'UNHATCHED'))
+            if len(all_unhatched_seeds) == 0:
+                process_logger.info("No more experiments to run. Terminating.")
+                break
 
-            args = load_config_from_json(str(seed_dir / 'args.json'))
-            dir_manager = DirectoryManager.init_from_seed_path(seed_dir)
-            logger = create_logger(
-                save_dir=dir_manager.seed_dir,
-                streamHandle=False
-                )
+            else:
+                # gets the next experiment path
+                seed_dir = all_unhatched_seeds[0]
+                os.remove(str(seed_dir / 'UNHATCHED'))
 
-            pbar = tqdm(position=process_i + (1 + serial_args.n_processes) * call_i)
-            train(args, dir_manager, logger, pbar)
+                args = load_config_from_json(str(seed_dir / 'args.json'))
+                dir_manager = DirectoryManager.init_from_seed_path(seed_dir)
+                logger = create_logger(save_dir=dir_manager.seed_dir, streamHandle=False)
 
-            open(str(seed_dir / 'COMPLETED'), 'w+').close()
+                if serial_args.pbar:
+                    pbar = tqdm(position=process_i + (1 + serial_args.n_processes) * call_i)
+                else:
+                    pbar = None
+                    process_logger.info(f'Starting {dir_manager.storage_dir.name}/{dir_manager.experiment_dir.name}/{dir_manager.seed_dir.name}')
+
+                train(args, dir_manager, logger, pbar)
+
+                if not serial_args.pbar:
+                    process_logger.info(f'COMPLETED {dir_manager.storage_dir.name}/{dir_manager.experiment_dir.name}/{dir_manager.seed_dir.name}\n')
+                open(str(seed_dir / 'COMPLETED'), 'w+').close()
 
         except Exception as e:
+            if not serial_args.pbar:
+                process_logger.info(f'CRASHED in {seed_dir.parent.parent.name}/{seed_dir.parent.name}/{seed_dir.name}\n')
             with open(str(seed_dir / 'CRASH.txt'), 'w+') as f:
                 f.write(f'Crashed at : {datetime.datetime.now().strftime("%Y/%m/%d %H:%M:%S")}')
                 f.write(str(e))
@@ -69,7 +86,8 @@ def run_schedule(serial_args, process_i=0):
             create_comparative_figure(storage_dir, create_logger(streamHandle=False))
 
         except Exception as e:
-            print(f"WARNING: Process {process_i} was not able to create comparative graphs: '{e}'")
+            if not serial_args.pbar:
+                process_logger.info(f"WARNING: was not able to create comparative graphs")
             continue
 
 if __name__ == '__main__':
